Moved DashboardData type to lib/api/user.ts and added source and exportPath properties.
Relocated the DashboardData interface from types/dashboard.ts to api/user.ts to establish it as the proper source. Added the required source and exportPath properties to the interface, updated the mock data, and fixed all imports in the application to use the new export path.

Closes #805
